pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptdefender"
version = "0.1.0"
description = "Explainable prompt injection detector with evals + policy gating"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "pydantic>=2.7",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "PyYAML>=6.0",
  "pytest>=8.0"
]

[project.optional-dependencies]
embeddings = [
  "sentence-transformers>=2.2.2,<3.0"
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = "."

[tool.setuptools.packages.find]
include = ["promptdefender*"]
exclude = ["app*", "benchmarks*", "tests*"]

[tool.setuptools.package-data]
promptdefender = ["heuristics.yaml", "assets/*.txt"]
